+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - `Request` and `Response` `content_type` properties now return `MediaType`
   instead of a string. Use `str(request.content_type)` for former behaviour.
 
+### Enhancements
+
+- `Request` contains `text()` method for decoding a request body.
+
 ## 0.10.0 (2024/07/14)
 
 ### Enhancements